moved ActiveHidSmDevice to ReadingMethod folder
