+++ Source/SpaceMouseReader/Private/ReadingMethod/ActiveHidSmDevice.cpp
@@ -1,7 +1,7 @@
 ﻿// Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "ActiveHidSmDevice.h"
+#include "ReadingMethod/ActiveHidSmDevice.h"
 
 #include "hidapi.h"
